feat(user): 添加第三方登录功能- 新增 OAuth 相关的接口和服务实现 - 增加 Gitee 第三方登录支持 - 实现用户信息获取和登录日志记录 - 优化国际化支持，更新相关错误信息

+++ mall-core/src/main/java/com/jiang/mall/domain/entity/Order.java
@@ -72,12 +72,12 @@ public class Order implements Serializable {
     /**
      * 订单状态
      */
-    private Byte status;
+    private Integer status;
 
     /**
      * 支付方式
      */
-    private Byte paymentMethod;
+    private  Integer paymentMethod;
 
     /**
      * 创建时间，自动填充
